chore: `config.hpp` cleanup

+++ src/config.hpp
@@ -12,7 +12,8 @@ inline constexpr unsigned int COLUMN_CNT = 2;
 inline constexpr unsigned int POINTS_CNT = COLUMN_CNT * 4;
 inline constexpr float OUT_WID = 1280.0f;
 inline constexpr float OUT_HEI = 720.0f;
-inline constexpr double TOTAL_AREA = OUT_WID * OUT_HEI;
+inline constexpr int TOTAL_AREA =
+    OUT_WID * OUT_HEI; // only used internally to calc frame area% consts
 inline constexpr unsigned int MAX_RETRIES = 30;
 
 inline constexpr bool SHOW_RAW = true;
@@ -21,7 +22,8 @@ inline constexpr bool SHOW_GUI = true;
 
 inline constexpr int MOTION_THRESH_INTENSITY =
     30; // px intens diff to count as 'changed'
-inline constexpr int CONTENT_THRESH_INTENSITY = MOTION_THRESH_INTENSITY;
+inline constexpr int CONTENT_THRESH_INTENSITY =
+    MOTION_THRESH_INTENSITY; // kept as separate for potential fine-tuning
 inline constexpr int MOTION_TRIGGER_PXS =
     1000; // how many pxs must change to trigger movement
 inline constexpr int SLIDE_TRIGGER_PXS = static_cast<int>(TOTAL_AREA * 0.05);
